+++ packages/framework/test/fixtures/app-base-starter/imports.ts
@@ -1,4 +1,0 @@
-export { MainConfiguration as Configuration } from './configuration';
-export * as defaultConfig from './config/config.default';
-export { SystemController } from './controller/system.controller';
-export { MainWindow } from './window/main.window';